Isolate findIconByPageColor function to separate file
