+++ next/components/molecules/SearchPageNew/SearchCardComposed.tsx
@@ -1,15 +1,3 @@
-import PageBlueIcon from '@assets/images/page-blue-icon.svg'
-import PageBlueIconSmall from '@assets/images/page-blue-icon-small.svg'
-import PageBrownIcon from '@assets/images/page-brown-icon.svg'
-import PageBrownIconSmall from '@assets/images/page-brown-icon-small.svg'
-import PageGreenIcon from '@assets/images/page-green-icon.svg'
-import PageGreenIconSmall from '@assets/images/page-green-icon-small.svg'
-import PagePurpleIcon from '@assets/images/page-purple-icon.svg'
-import PagePurpleIconSmall from '@assets/images/page-purple-icon-small.svg'
-import PageRedIcon from '@assets/images/page-red-icon.svg'
-import PageRedIconSmall from '@assets/images/page-red-icon-small.svg'
-import PageYellowIcon from '@assets/images/page-yellow-icon.svg'
-import PageYellowIconSmall from '@assets/images/page-yellow-icon-small.svg'
 import { ChevronRightIcon } from '@assets/ui-icons'
 import { Enum_Page_Pagecolor, Enum_Pagecategory_Color } from '@backend/graphql'
 import ImagePlaceholder from '@components/atoms/ImagePlaceholder'
@@ -19,6 +7,7 @@ import { SearchResult } from '@components/molecules/SearchPageNew/searchDataFetc
 import { getCategoryColorLocalStyle } from '@utils/colors'
 import { generateImageSizes } from '@utils/generateImageSizes'
 import { isDefined } from '@utils/isDefined'
+import { findIconByPageColor } from '@utils/pageIcons'
 import Image from 'next/image'
 import React, { ReactNode } from 'react'
 import { twMerge } from 'tailwind-merge'
@@ -189,19 +178,6 @@ SearchCardComposed.Button = function ({ className }: { className?: string }) {
       <ChevronRightIcon />
     </div>
   )
-}
-
-const findIconByPageColor = (pageColor: Enum_Pagecategory_Color) => {
-  const icons = {
-    red: { default: PageRedIcon, small: PageRedIconSmall },
-    blue: { default: PageBlueIcon, small: PageBlueIconSmall },
-    green: { default: PageGreenIcon, small: PageGreenIconSmall },
-    yellow: { default: PageYellowIcon, small: PageYellowIconSmall },
-    purple: { default: PagePurpleIcon, small: PagePurpleIconSmall },
-    brown: { default: PageBrownIcon, small: PageBrownIconSmall },
-  }
-
-  return icons[pageColor] ?? icons.red
 }
 
 export default SearchCardComposed
